+++ tests/Response/Types/PagerfantaTypeTest.php
@@ -29,7 +29,7 @@ class PagerfantaTypeTest extends TestCase
         $this->assertIsArray($obj->getIncludes());
         $this->assertIsArray($obj->getMeta());
 
-        $obj->withKey('bob')->withIncludes('foo', 'bar')->withMeta($meta = ['meta' => 'bob']);
+        $obj->key('bob')->include('foo', 'bar')->meta($meta = ['meta' => 'bob']);
 
         $this->assertEquals(static::class, $obj->getTransformer());
         $this->assertEquals('bob', $obj->getKey());
@@ -37,5 +37,19 @@ class PagerfantaTypeTest extends TestCase
         $this->assertEquals(['foo', 'bar'], $obj->getIncludes());
         $this->assertEquals($meta, $obj->getMeta());
         $this->assertInstanceOf(Collection::class, $obj->asResource());
+    }
+
+    /**
+     * @group services
+     * @group services-response
+     * @group services-response-type
+     */
+    public function testCanChangeUrl()
+    {
+        $obj = new PagerfantaType(new Pagerfanta(new ArrayAdapter([])), static::class, 'http://www.example.com');
+
+        $obj->url('https://example.com');
+
+        $this->assertEquals('https://example.com', $obj->getUrl());
     }
 }